refactored, added "file_size" field to task["outputs"]
